Added much more functionality

+++ src/main/java/pl/vmrent/web/controller/machine/MachineListController.java
@@ -2,6 +2,7 @@ package pl.vmrent.web.controller.machine;
 
 import pl.vmrent.web.model.machine.Machine;
 import pl.vmrent.web.service.MachineService;
+import pl.vmrent.web.service.RentService;
 
 import javax.annotation.PostConstruct;
 import javax.faces.view.ViewScoped;
@@ -17,6 +18,9 @@ public class MachineListController implements Serializable
     @Inject
     private MachineService machineService;
 
+    @Inject
+    private RentService rentService;
+
     private List<Machine> machines;
 
     @PostConstruct
@@ -27,8 +31,17 @@ public class MachineListController implements Serializable
 
     public String deleteMachine(Machine machine)
     {
-        machineService.deleteMachine(machine);
-        return "show_vms.xhtml?faces-redirect=true";
+        if (!rentService.isMachineRented(machine))
+        {
+            machineService.deleteMachine(machine);
+            return "show_vms.xhtml?faces-redirect=true";
+        }
+        return null;
+    }
+
+    public boolean isRented(Machine machine)
+    {
+        return rentService.isMachineRented(machine);
     }
 
     public List<Machine> getMachines()
